Improve docstring parser
Using split may lead to some unpredicted parsing. Use regex to improve the
parser in order to avoid that.

For example:

    """Some content before tokens

    More content, but with : more content

    @token1: value1

    @token2: value2
    """

Using the old parser and the docstring above that will generate a invalid
token like the following:

    Some content before tokens
    More content, but with :
      more content

In the above snippet all content up to `:` is the token and all content after
`:` is the value, which is not valid since the it does not start with the `@`.

On the other hand the new parser gets just the expected tokens.